@@ -1,12 +1,16 @@
-module github.com/h2non/imaginary
+module github.com/jgibo/imaginary
 
-go 1.12
+go 1.20
 
 require (
-	github.com/garyburd/redigo v1.6.0 // indirect
-	github.com/h2non/bimg v1.1.7
-	github.com/h2non/filetype v1.1.0
+	github.com/h2non/bimg v1.1.9
+	github.com/h2non/filetype v1.1.3
+	github.com/rs/cors v1.9.0
+	gopkg.in/throttled/throttled.v2 v2.2.5
+)
+
+require (
+	github.com/gomodule/redigo v1.8.9 // indirect
 	github.com/hashicorp/golang-lru v0.0.0-20160813221303-0a025b7e63ad // indirect
-	github.com/rs/cors v0.0.0-20170727213201-7af7a1e09ba3
-	gopkg.in/throttled/throttled.v2 v2.0.3
+	github.com/throttled/throttled v2.2.5+incompatible // indirect
 )
