@@ -1,8 +1,8 @@
 package main
 
 import (
-	"io/ioutil"
 	"net/http"
+	"os"
 	"path"
 	"strings"
 )
@@ -44,7 +44,7 @@ func (s *FileSystemImageSource) buildPath(file string) (string, error) {
 }
 
 func (s *FileSystemImageSource) read(file string) ([]byte, error) {
-	buf, err := ioutil.ReadFile(file)
+	buf, err := os.ReadFile(file)
 	if err != nil {
 		return nil, ErrInvalidFilePath
 	}
