@@ -147,7 +147,7 @@ func setCacheHeaders(next http.Handler, ttl int) http.Handler {
 		ttlDiff := time.Duration(ttl) * time.Second
 		expires := time.Now().Add(ttlDiff)
 
-		w.Header().Add("Expires", strings.Replace(expires.Format(time.RFC1123), "UTC", "GMT", -1))
+		w.Header().Add("Expires", strings.ReplaceAll(expires.Format(time.RFC1123), "UTC", "GMT"))
 		w.Header().Add("Cache-Control", getCacheControl(ttl))
 	})
 }
